requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blis==1.0.1
Bottleneck==1.3.7
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
deepdiff==8.0.1
deepl==1.20.0
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
elevenlabs==1.12.1
emoji==2.14.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.1.0
faiss-cpu==1.9.0
fastai==2.7.18
fastapi==0.115.4
fastcore==1.7.19
fastdownload==0.0.7
fastjsonschema==2.20.0
fastprogress==1.0.3
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwipiepy==0.20.2
kiwipiepy_model==0.20.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.3
langchain-community==0.2.16
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-teddynote==0.3.30
langchain-text-splitters==0.3.2
langchain-unstructured==0.1.5
langcodes==3.4.1
langdetect==1.0.9
langgraph==0.2.53
langgraph-checkpoint==2.0.5
langgraph-sdk==0.1.36
langsmith==0.1.144
language_data==1.2.0
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.0
mkl_fft==1.3.10
mkl_random==1.2.7
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.14.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numexpr==2.8.7
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.54.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orderly-set==5.2.2
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer==20191125
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
platformdirs==4.3.6
plotly==5.24.1
posthog==3.7.2
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.14
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.26.1
Send2Trash==1.8.3
setuptools==75.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.2
streamlit==1.40.1
streamlit-chat==0.1.1
sympy==1.13.1
tavily-python==0.5.0
tenacity==8.5.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
torchtext==0.17.2
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.1
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
types-python-dateutil==2.9.0.20240906
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.3.27
ultralytics-thop==2.0.10
unstructured==0.16.6
unstructured-client==0.25.9
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wget==3.2
wheel==0.44.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.1
yarl==1.17.1
zipp==3.21.0